+++ modules/apps/frontend-data-set/frontend-data-set-web/src/main/resources/META-INF/resources/actions/types/index.js
